+++ packages/subscription-controller/src/index.ts
@@ -60,6 +60,7 @@ export {
   PRODUCT_TYPES,
   RECURRING_INTERVALS,
   PAYMENT_TYPES,
+  SubscriptionUserEvent,
 } from './types';
 export { SubscriptionServiceError } from './errors';
 export { Env, SubscriptionControllerErrorMessage } from './constants';